Se actualiza servicio para eliminar archivo de la cola de tareas y de la carpeta fisica.

+++ conversion_app/vistas/vistas.py
@@ -1,4 +1,5 @@
 import logging
+import os
 from flask import request
 from flask_jwt_extended import jwt_required, create_access_token, get_jwt_identity
 from flask_restful import Resource
@@ -8,14 +9,14 @@ from celery import Celery
 import datetime
 
 from app import db
-from modelos import Usuario, Task, TaskSchema, File
+from modelos import Usuario, Task, TaskSchema, File, Status
 
 
 task_schema = TaskSchema()
 
 celery = Celery(__name__, broker='redis://redis:6379/0')
 
-def esCorreoElectronico(texto):
+def check_email(texto):
     # Definimos una expresión regular para validar direcciones de correo electrónico
     patron = r"^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$"
     
@@ -62,7 +63,7 @@ class VistaLogIn(Resource):
         usuario = None
         contrasena_encriptada = hashlib.md5(request.json["password"].encode('utf-8')).hexdigest()
 
-        if esCorreoElectronico(data.get('usuario')):
+        if check_email(data.get('usuario')):
             usuario = Usuario.query.filter(Usuario.email == data["usuario"],
                                        Usuario.password == contrasena_encriptada).first()
         else:
@@ -81,7 +82,7 @@ class VistaTasks(Resource):
     @jwt_required()
     def get(self):
         current_user = get_jwt_identity()
-        tareas = Task.query.filter(id=current_user)
+        tareas = Task.query.filter_by(userId=current_user).all()
         # Aplica parámetros de consulta, si están presentes
         max_param = request.args.get('max', type=int)
         order_param = request.args.get('order', type=int)
@@ -103,18 +104,18 @@ class VistaTasks(Resource):
     @jwt_required()
     def post(self):
         current_user = get_jwt_identity()
-        fileName = request.json["fileName"]
-        originalFormat = obtener_formato_archivo(fileName)
-        newFormat = request.json["newFormat"]
+        file_name = request.json["fileName"]
+        original_format = obtener_formato_archivo(file_name)
+        new_format = request.json["newFormat"]
         status="UPLOADED"
-        uploadTime = datetime.datetime.now()
-        new_file = File(fileName=fileName,originalFormat=originalFormat,newFormat=newFormat)
+        upload_time = datetime.datetime.now()
+        new_file = File(fileName=file_name,originalFormat=original_format,newFormat=new_format)
         db.session.add(new_file)
         db.session.commit()
-        new_task = Task(idFile=new_file.id,status=status,uploadTime=uploadTime,userId=current_user)
+        new_task = Task(idFile=new_file.id,status=status,uploadTime=upload_time,userId=current_user)
         db.session.add(new_task)
         db.session.commit()
-        send_task_to_process.apply_async(args=[new_task.id,fileName,newFormat], queue="process_task_converter")
+        send_task_to_process.apply_async(args=[new_task.id,file_name,new_format,new_task.userId], queue="process_task_converter")
 
         return "Su transaccion esta en proceso con el task_id: {}".format(new_task.id)
 
@@ -126,13 +127,31 @@ class VistaTask(Resource):
     @jwt_required()
     def delete(self, id_task):
         registro = Task.query.get_or_404(id_task)
-        db.session.delete(registro)
-        db.session.commit()
-        return "Registro eliminado exitosamente"
+        file = File.query.get_or_404(registro.idFile)
+
+        if registro.status == Status.PROCESSED:
+            output_filename = os.path.splitext(file.fileName)[0]+ f'_{registro.id}' + f'_{registro.userId}' + f'.{file.newFormat}'
+            output_video_path = "/app/files/converted/" + output_filename.split("/")[-1]
+            
+            if os.path.exists(output_video_path):
+                os.remove(output_video_path)
+                
+            original_filename = os.path.splitext(file.fileName)[0]+ f'_{registro.id}' + f'_{registro.userId}' + f'.{file.originalFormat}'
+            original_video_path = "/app/files/original/" + original_filename.split("/")[-1]
+            
+            if os.path.exists(original_video_path):
+                os.remove(original_video_path)
+
+            db.session.delete(registro)
+            db.session.delete(file)
+            db.session.commit()
+            
+            return "Registro eliminado exitosamente "+ f'_{output_video_path}'+  f'_{original_video_path}'
+        else:
+            return "La tarea no puede ser eliminada. Tarea en proceso"+ f'_{output_video_path}'+  f'_{original_video_path}'
     
     
 
-            
 
 @celery.task(name="convert_process")
 def send_task_to_process(*args):
